Rename some files and components.
