+++ src/app/loadout/loadout-ui/LoadoutItemCategorySection.m.scss
